Flytter utledPrioritet til en util klasse
