+++ src/main/kotlin/no/nav/familie/ef/mottak/service/AutomatiskJournalføringService.kt
@@ -2,6 +2,7 @@ package no.nav.familie.ef.mottak.service
 
 import no.nav.familie.ef.mottak.integration.SaksbehandlingClient
 import no.nav.familie.ef.mottak.repository.domain.Søknad
+import no.nav.familie.ef.mottak.util.UtledPrioritetForSøknadUtil
 import no.nav.familie.kontrakter.ef.journalføring.AutomatiskJournalføringRequest
 import no.nav.familie.kontrakter.ef.journalføring.AutomatiskJournalføringResponse
 import no.nav.familie.kontrakter.felles.ef.StønadType
@@ -15,7 +16,6 @@ class AutomatiskJournalføringService(
     val saksbehandlingClient: SaksbehandlingClient,
     val søknadService: SøknadService,
     val taskService: TaskService,
-    val oppgaveService: OppgaveService,
 ) {
 
     val logger: Logger = LoggerFactory.getLogger(this::class.java)
@@ -33,7 +33,7 @@ class AutomatiskJournalføringService(
             journalpostId = journalpostId,
             stønadstype = stønadstype,
             mappeId = mappeId,
-            prioritet = oppgaveService.utledPrioritet(søknad),
+            prioritet = UtledPrioritetForSøknadUtil.utledPrioritet(søknad),
         )
         try {
             val respons =
